example/pyproject.toml
[project]
name = "ft-icon-example"
version = "0.1.0"
description = "Example FastHTML app using ft-icon"
requires-python = ">=3.11"
dependencies = [
    "python-fasthtml>=0.1.0",
    "ft-icon"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/example"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
ft-icon = { git = "https://github.com/banditburai/ft-icon.git" }